network: refactor package structure
